requirements.txt
pynacl
requests
python-dotenv
openai
langchain
langchain-openai
langchain-community
langchain-pinecone
pinecone-client